+++ packages/styled/react/src/StyledProvider.tsx
@@ -5,6 +5,7 @@ import { propertyTokenMap } from './propertyTokenMap';
 import type { COLORMODES } from './types';
 import {
   convertTokensToCssVariables,
+  generateMergedThemeTokens,
   platformSpecificSpaceUnits,
 } from './utils';
 import { createGlobalStylesWeb } from './createGlobalStylesWeb';
@@ -45,6 +46,7 @@ const setCurrentColorMode = (inputColorMode: string | undefined) => {
   //   colorModeSet = true;
   // }
 };
+
 export const StyledProvider: React.FC<{
   config: Config;
   colorMode?: COLORMODES;
@@ -63,6 +65,8 @@ export const StyledProvider: React.FC<{
   });
 
   const { themes } = useTheme();
+  const styledContext = useStyled();
+  const isParentProviderExist = styledContext?.config ? true : false;
 
   const themeContextValue = React.useMemo(() => {
     if (colorMode) {
@@ -95,6 +99,10 @@ export const StyledProvider: React.FC<{
         configWithPlatformSpecificUnits
       );
     }
+
+    configWithPlatformSpecificUnits = generateMergedThemeTokens(
+      configWithPlatformSpecificUnits
+    );
 
     return configWithPlatformSpecificUnits;
   }, [config]);
@@ -130,7 +138,9 @@ export const StyledProvider: React.FC<{
       documentElement.classList.add(`gs`);
 
       if (currentColorMode) {
-        document.body.setAttribute('data-theme-id', currentColorMode);
+        if (!isParentProviderExist) {
+          document.body.setAttribute('data-theme-id', currentColorMode);
+        }
         documentElement.classList.add(`gs-${currentColorMode}`);
       } else {
         documentElement.classList.add(`gs-light`);
@@ -145,10 +155,14 @@ export const StyledProvider: React.FC<{
         if (Platform.OS === 'web') {
           if (currentColor) {
             if (currentColor === 'dark') {
-              document.body.setAttribute('data-theme-id', 'dark');
+              if (!isParentProviderExist) {
+                document.body.setAttribute('data-theme-id', 'dark');
+              }
               documentElement.classList.remove(`gs-light`);
             } else {
-              document.body.setAttribute('data-theme-id', 'light');
+              if (!isParentProviderExist) {
+                document.body.setAttribute('data-theme-id', 'light');
+              }
               documentElement.classList.remove(`gs-dark`);
             }
             documentElement.classList.add(`gs-${currentColor}`);