+++ packages/styled/react/src/resolver/StyledValueToCSSObject.ts
@@ -1,6 +1,5 @@
 import type { CSSObject, StyledValue } from '../types';
-import { deepMerge, resolvedTokenization } from '../utils';
-import { deepClone } from '../utils/cssify/utils/common';
+import { resolvedTokenization } from '../utils';
 
 export function StyledValueToCSSObject(
   input: StyledValue | undefined,
@@ -11,6 +10,7 @@ export function StyledValueToCSSObject(
   if (!input) {
     return {};
   }
+
   return resolvedTokenization(input, CONFIG, ignoreKeys, deleteIfTokenNotExist);
 }
 export function themeStyledValueToCSSObject(
@@ -20,34 +20,36 @@ export function themeStyledValueToCSSObject(
 ) {
   let themeResolved1: any = {};
   if (CONFIG?.themes) {
-    const tokens = deepClone(CONFIG.tokens);
-    Object.keys(CONFIG?.themes).forEach((key: any) => {
-      const themeTokens = CONFIG?.themes[key];
-      Object.keys(themeTokens).forEach((tokenKey1: any) => {
-        Object.keys(themeTokens[tokenKey1]).forEach((tokenKey: any) => {
-          delete tokens[tokenKey1][tokenKey];
-        });
-      });
-    });
+    // const tokens = deepClone(CONFIG.tokens);
+    // Object.keys(CONFIG?.themes).forEach((key: any) => {
+    //   const themeTokens = CONFIG?.themes[key];
+    //   Object.keys(themeTokens).forEach((tokenKey1: any) => {
+    //     Object.keys(themeTokens[tokenKey1]).forEach((tokenKey: any) => {
+    //       delete tokens[tokenKey1][tokenKey];
+    //     });
+    //   });
+    // });
 
     // debugger;
 
-    Object.keys(CONFIG?.themes).forEach((key: any) => {
-      const themeResolved = StyledValueToCSSObject(
-        input,
-        {
-          ...CONFIG,
-          tokens: deepMerge(deepClone(tokens), CONFIG.themes[key]),
-        },
-        ignoreKeys,
-        true
-      );
+    Object.keys(CONFIG?.themes).forEach((themeName: any) => {
+      if (themeName !== 'tokens') {
+        const themeResolved = StyledValueToCSSObject(
+          input,
+          {
+            ...CONFIG,
+            tokens: CONFIG?.themes?.tokens[themeName],
+          },
+          ignoreKeys,
+          true
+        );
 
-      Object.keys(themeResolved).forEach((key) =>
-        themeResolved[key] === undefined ? delete themeResolved[key] : {}
-      );
+        Object.keys(themeResolved).forEach((key: any) =>
+          themeResolved[key] === undefined ? delete themeResolved[key] : {}
+        );
 
-      themeResolved1[key] = themeResolved;
+        themeResolved1[themeName] = themeResolved;
+      }
     });
   }